+++ components/regions/index.js
@@ -4,9 +4,10 @@ import FaList from 'react-icons/lib/fa/list'
 import Section from '../section'
 import Summary from '../summary'
 import TechnicalDoc from '../technical-doc'
-import ByName from '../by-name'
 
-import RegionsList from './examples/regions-list'
+import ByName from '../demo/by-name'
+
+import RegionsList from './regions-list'
 
 import doc from './doc'
